Move `<Attacks />` back to be demo local; Replace any's with placeholder types.
