+++ ui/src/components/index.ts
@@ -5,4 +5,3 @@ export * from './Interpret';
 export * from './Interpreters';
 export * from './TaskCards';
 export * from './Attack';
-export * from './Attacks';
